+++ src/main/java/net/silentchaos512/gems/SideProxy.java
@@ -21,16 +21,17 @@ import net.silentchaos512.gems.command.ChaosCommand;
 import net.silentchaos512.gems.command.HungryCommand;
 import net.silentchaos512.gems.compat.gear.SGearProxy;
 import net.silentchaos512.gems.compat.gear.SGearStatHandler;
+import net.silentchaos512.gems.config.GemsConfig;
 import net.silentchaos512.gems.crafting.tokenenchanter.TokenEnchanterRecipeManager;
 import net.silentchaos512.gems.event.TraitEvents;
-import net.silentchaos512.gems.lib.soul.GearSoulPart;
-import net.silentchaos512.gems.network.Network;
-import net.silentchaos512.gems.util.SoulEvents;
 import net.silentchaos512.gems.init.*;
 import net.silentchaos512.gems.item.TeleporterLinker;
 import net.silentchaos512.gems.lib.ColorHandlers;
 import net.silentchaos512.gems.lib.chaosbuff.ChaosBuffManager;
+import net.silentchaos512.gems.lib.soul.GearSoulPart;
 import net.silentchaos512.gems.lib.soul.Soul;
+import net.silentchaos512.gems.network.Network;
+import net.silentchaos512.gems.util.SoulEvents;
 import net.silentchaos512.gems.util.gen.GenModels;
 import net.silentchaos512.gems.util.gen.GenRecipes;
 import net.silentchaos512.gems.world.GemsWorldFeatures;
@@ -71,6 +72,7 @@ class SideProxy {
 
         registerContainersCommon();
 
+        GemsConfig.init();
         Network.init();
 
         if (SilentGems.isDevBuild()) {