+++ src/main/java/net/silentchaos512/gems/SilentGems.java
@@ -15,8 +15,8 @@ import java.util.Random;
 public final class SilentGems {
     public static final String MOD_ID = "silentgems";
     public static final String MOD_NAME = "Silent's Gems";
-    public static final String VERSION = "3.0.9";
-    public static final boolean RUN_GENERATORS = true;
+    public static final String VERSION = "3.0.10";
+    public static final boolean RUN_GENERATORS = false;
     public static final String RESOURCE_PREFIX = MOD_ID + ":";
 
     public static final Random random = new Random();