+++ src/main/java/net/silentchaos512/gems/client/gui/DebugOverlay.java
@@ -2,6 +2,7 @@ package net.silentchaos512.gems.client.gui;
 
 import com.google.common.collect.ImmutableList;
 import net.silentchaos512.gems.client.ClientPlayerInfo;
+import net.silentchaos512.gems.config.GemsConfig;
 import net.silentchaos512.lib.client.gui.DebugRenderOverlay;
 
 import javax.annotation.Nonnull;
@@ -12,8 +13,8 @@ public class DebugOverlay extends DebugRenderOverlay {
     @Override
     public List<String> getDebugText() {
         return ImmutableList.of(
-                "Chaos (world)=" + ClientPlayerInfo.playerChaos,
-                "Chaos (player)=" + ClientPlayerInfo.worldChaos
+                "Chaos (player)=" + ClientPlayerInfo.playerChaos,
+                "Chaos (world)=" + ClientPlayerInfo.worldChaos
         );
     }
 
@@ -24,6 +25,6 @@ public class DebugOverlay extends DebugRenderOverlay {
 
     @Override
     public boolean isHidden() {
-        return false;
+        return !(GemsConfig.COMMON.debugMasterSwitch.get() && GemsConfig.COMMON.debugShowOverlay.get());
     }
 }